+++ types/payments/index.d.ts
@@ -1,4 +1,3 @@
-/// <reference types="node" />
 import { Network } from '../networks';
 import { p2data as embed } from './embed';
 import { p2ms } from './p2ms';
@@ -24,6 +23,7 @@ export interface Payment {
     redeem?: Payment;
     witness?: Buffer[];
 }
+export declare type PaymentCreator = (a: Payment, opts?: PaymentOpts) => Payment;
 export declare type PaymentFunction = () => Payment;
 export interface PaymentOpts {
     validate?: boolean;